Promoted cpu_Info.Cores to uint64 to avoid struct padding

+++ kernel/cpu.c
@@ -3,7 +3,7 @@
 
 cpu_Info cpu_GetInfo() {
 	cpu_Info info = {
-		.Cores = *(uint16*)0x5012,
+		.Cores = (uint64)*(uint16*)0x5012,
 		.Frequency = ((uint64)1000000) * (uint64)(*(uint16*)0x5010), // MHz to Hz.
 		.Memory = (((uint64)1)<<20) * (uint64)(*(uint32*)0x5020), // MiB to B.
 	};
