+++ kernel/cpu.h
@@ -3,7 +3,7 @@
 #define CPU_H
 
 typedef struct {
-	uint16 Cores;
+	uint64 Cores;
 	uint64 Frequency;
 	uint64 Memory;
 } cpu_Info;
